research/screen-page-classification/requirements.txt
pytorch-lightning==2.3.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=9.5.0
pyyaml>=6.0.1
